+++ R/predict.R
@@ -1,81 +1,50 @@
-################################################################################
 ##
-##   R package splines2 by Wenjie Wang and Jun Yan
-##   Copyright (C) 2016-2018
+## R package splines2 by Wenjie Wang and Jun Yan
+## Copyright (C) 2016-2020
 ##
-##   This file is part of the R package splines2.
+## This file is part of the R package splines2.
 ##
-##   The R package splines2 is free software: You can redistribute it and/or
-##   modify it under the terms of the GNU General Public License as published
-##   by the Free Software Foundation, either version 3 of the License, or
-##   any later version (at your option). See the GNU General Public License
-##   at <http://www.gnu.org/licenses/> for details.
+## The R package splines2 is free software: You can redistribute it and/or
+## modify it under the terms of the GNU General Public License as published by
+## the Free Software Foundation, either version 3 of the License, or any later
+## version (at your option). See the GNU General Public License at
+## <https://www.gnu.org/licenses/> for details.
 ##
-##   The R package splines2 is distributed in the hope that it will be useful,
-##   but WITHOUT ANY WARRANTY without even the implied warranty of
-##   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+## The R package splines2 is distributed in the hope that it will be useful,
+## but WITHOUT ANY WARRANTY without even the implied warranty of
+## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 ##
-################################################################################
 
-
-##' Evaluate a Spline Basis
+##' Evaluate a Spline Basis at specified points
 ##'
-##' This function evaluates a predefined spline basis at (new) given values.
+##' This function evaluates a predefined spline basis at a (new) given \code{x}.
 ##'
 ##' These are methods for the generic function \code{predict} for objects
 ##' inheriting from class \code{bSpline2}, \code{ibs}, \code{mSpline},
-##' \code{iSpline}, or \code{cSpline}.  If \code{newx} is not given, the
-##' function returns the input object.  For object returned by function
-##' \code{\link{cSpline}}, the \code{mSpline} and \code{iSpline} objects shipped
-##' in attributes should not be evaluated by this function if \code{rescale} is
-##' \code{TRUE}.  See \code{\link{cSpline}} for details.
+##' \code{iSpline}, \code{cSpline}, or \code{bernsteinPoly}.  If \code{newx} is
+##' not given, the function returns the input object.
 ##'
 ##' @name predict
 ##' @param object Objects of class \code{bSpline2}, \code{ibs}, \code{mSpline},
-##'     \code{iSpline}, or \code{cSpline} having attributes describing
-##'     \code{knots}, \code{degree}, etc.
+##'     \code{iSpline}, \code{cSpline}, or \code{bernsteinPoly} with attributes
+##'     describing \code{knots}, \code{degree}, etc.
 ##' @param newx The \code{x} values at which evaluations are required.
-##' @param ... Optional argument for future usage.
+##' @param ... Optional argument that are not used.
 ##'
-##' @return An object just like the \code{object} input, except evaluated at
+##' @return
+##' An object just like the \code{object} input, except evaluated at
 ##' the new values of \code{x}.
 ##'
-##' @examples
-##' library(splines2)
-##' x <- seq.int(0, 1, 0.2)
-##' knots <- c(0.3, 0.5, 0.6)
-##' newX <- seq.int(0.1, 0.9, 0.2)
+##' @example inst/examples/ex-predict.R
 ##'
-##' ## for B-splines
-##' bsMat <- bSpline(x, knots = knots, degree = 2)
-##' predict(bsMat, newX)
-##'
-##' ## for integral of B-splines
-##' ibsMat <- ibs(x, knots = knots, degree = 2)
-##' predict(ibsMat, newX)
-##'
-##' ## for derivative of B-splines
-##' dbsMat <- dbs(x, knots = knots, degree = 2)
-##' predict(dbsMat, newX)
-##'
-##' ## for M-spline
-##' msMat <- mSpline(x, knots = knots, degree = 2)
-##' predict(msMat, newX)
-##'
-##' ## for I-spline
-##' isMat <- iSpline(x, knots = knots, degree = 2)
-##' predict(isMat, newX)
-##'
-##' ## for C-spline
-##' csMat <- cSpline(x, knots = knots, degree = 2)
-##' predict(csMat, newX)
 ##' @seealso
 ##' \code{\link{bSpline}} for B-splines;
-##' \code{\link{ibs}} for integral of B-splines;
-##' \code{\link{dbs}} for derivative of B-splines;
+##' \code{\link{ibs}} for integrals of B-splines;
+##' \code{\link{dbs}} for derivatives of B-splines;
 ##' \code{\link{mSpline}} for M-splines;
 ##' \code{\link{iSpline}} for I-splines;
 ##' \code{\link{cSpline}} for C-splines.
+##'
 ##' @importFrom stats predict
 NULL
 
@@ -155,4 +124,17 @@ predict.cSpline <- function(object, newx, ...)
            attributes(object)[c("degree", "knots", "Boundary.knots",
                                 "intercept", "scale")])
     do.call("cSpline", a)
+}
+
+
+##' @rdname predict
+##' @export
+predict.bernsteinPoly <- function(object, newx, ...)
+{
+    if (missing(newx))
+        return(object)
+    a <- c(list(x = newx),
+           attributes(object)[c("degree", "derivs", "integral",
+                                "intercept", "Boundary.knots")])
+    do.call("bernsteinPoly", a)
 }